oracle, testutils: seperated code based on review
