+++ oracle/midpoint_test.go
@@ -1,0 +1,1 @@
+package oracle_test
